refactoring anonymous class configurators - introduced InjectionConfigurator & BindingConfigurator
