+++ ee/src/main/java/org/jboss/as/ee/datasource/DataSourceDefinitionAnnotationParser.java
@@ -23,10 +23,9 @@
 package org.jboss.as.ee.datasource;
 
 import org.jboss.as.ee.component.Attachments;
+import org.jboss.as.ee.component.BindingConfigurator;
 import org.jboss.as.ee.component.BindingConfiguration;
-import org.jboss.as.ee.component.ClassConfigurator;
 import org.jboss.as.ee.component.EEApplicationClasses;
-import org.jboss.as.ee.component.EEModuleClassConfiguration;
 import org.jboss.as.ee.component.EEModuleClassDescription;
 import org.jboss.as.ee.component.EEModuleDescription;
 import org.jboss.as.server.deployment.DeploymentPhaseContext;
@@ -52,6 +51,7 @@ import java.util.List;
  *
  * @author John Bailey
  * @author Jason T. Greene
+ * @author <a href="mailto:ropalka@redhat.com">Richard Opalka</a>
  */
 public class DataSourceDefinitionAnnotationParser implements DeploymentUnitProcessor {
 
@@ -108,12 +108,7 @@ public class DataSourceDefinitionAnnotationParser implements DeploymentUnitProce
         final BindingConfiguration bindingConfiguration = this.getBindingConfiguration(datasourceDefinition);
         EEModuleClassDescription classDescription = applicationClasses.getOrAddClassByName(targetClass.name().toString());
         // add the binding configuration via a class configurator
-        classDescription.getConfigurators().add(new ClassConfigurator() {
-            @Override
-            public void configure(DeploymentPhaseContext context, EEModuleClassDescription description, EEModuleClassConfiguration configuration) throws DeploymentUnitProcessingException {
-                configuration.getBindingConfigurations().add(bindingConfiguration);
-            }
-        });
+        classDescription.getConfigurators().add(new BindingConfigurator(bindingConfiguration));
     }
 
     private BindingConfiguration getBindingConfiguration(final AnnotationInstance datasourceAnnotation) {
