+++ ejb3/src/main/java/org/jboss/as/ejb3/deployment/processors/EjbResourceInjectionAnnotationProcessor.java
@@ -25,12 +25,12 @@ package org.jboss.as.ejb3.deployment.processors;
 import static org.jboss.as.ee.component.Attachments.EE_APPLICATION_CLASSES_DESCRIPTION;
 import static org.jboss.as.server.deployment.Attachments.COMPOSITE_ANNOTATION_INDEX;
 
+import org.jboss.as.ee.component.BindingConfigurator;
 import org.jboss.as.ee.component.BindingConfiguration;
-import org.jboss.as.ee.component.ClassConfigurator;
 import org.jboss.as.ee.component.EEApplicationClasses;
-import org.jboss.as.ee.component.EEModuleClassConfiguration;
 import org.jboss.as.ee.component.EEModuleClassDescription;
 import org.jboss.as.ee.component.FieldInjectionTarget;
+import org.jboss.as.ee.component.InjectionConfigurator;
 import org.jboss.as.ee.component.InjectionSource;
 import org.jboss.as.ee.component.InjectionTarget;
 import org.jboss.as.ee.component.LookupInjectionSource;
@@ -183,14 +183,8 @@ public class EjbResourceInjectionAnnotationProcessor implements DeploymentUnitPr
         // TODO: class hierarchies? shared bindings?
         final EEApplicationClasses applicationClasses = unit.getAttachment(EE_APPLICATION_CLASSES_DESCRIPTION);
         final EEModuleClassDescription classDescription = applicationClasses.getOrAddClassByName(classInfo.name().toString());
-        classDescription.getConfigurators().add(new ClassConfigurator() {
-            public void configure(final DeploymentPhaseContext context, final EEModuleClassDescription description, final EEModuleClassConfiguration configuration) throws DeploymentUnitProcessingException {
-                configuration.getBindingConfigurations().add(bindingConfiguration);
-                if (injectionConfiguration != null) {
-                    configuration.getInjectionConfigurations().add(injectionConfiguration);
-                }
-            }
-        });
+        classDescription.getConfigurators().add(new BindingConfigurator(bindingConfiguration));
+        classDescription.getConfigurators().add(new InjectionConfigurator(injectionConfiguration));
     }
 
     private boolean isEmpty(final String string) {
